+++ slash/conf.py
@@ -25,6 +25,7 @@ config = Config({
         "last_failed_symlink": None // Doc("If set, specifies a symlink path to the last failed test log file"),
         "silence_loggers": [] // Doc("Logger names to silence"),
         "format": None // Doc("Format of the log line, as passed on to logbook. None will use the default format"),
+        "console_format": None // Doc("Optional format to be used for console output. Defaults to the regular format"),
         "localtime": False // Doc("Use local time for logging. If False, will use UTC"),
     },
     "run": {