+++ tests/test_console_log_truncation.py
@@ -10,20 +10,12 @@ from slash.log import ConsoleHandler
 def test_line_truncation(long_text, use_truncation, config_override):
 
     config_override('log.truncate_console_lines', use_truncation)
+    config_override('log.console_level', logbook.DEBUG)
 
     console = StringIO()
 
-
-    class _ConsoleHandler(ConsoleHandler):
-
-        @property
-        def stream(self):
-            return console
-
-    with slash.Session():
-
-        with _ConsoleHandler(level=logbook.DEBUG):
-            logbook.debug(long_text)
+    with slash.Session(console_stream=console):
+        logbook.debug(long_text)
 
     if use_truncation:
         assert long_text not in console.getvalue()